회원기능 만들기 : OAuth + session방식 사용하기

+++ forum/pages/api/post/delete.js
@@ -1,18 +1,27 @@
 import { connectDB } from "@/util/database";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(요청, 응답) {
-  const db = (await connectDB).db("forum");
-
   // 글 삭제 - ajax
   if (요청.method === "DELETE") {
     try {
-      let result = await db.collection("post").deleteOne({
-        _id: new ObjectId(요청.id),
-      });
-      console.log(result); // { acknowledged: true, deletedCount: 1 }
+      let session = await getServerSession(요청, 응답, authOptions);
+      const db = (await connectDB).db("forum");
 
-      응답.status(200).json("삭제완료");
+      // user의 author이 body 인 것을 찾음
+      let target = await db.collection('post').findOne({_id: new ObjectId(요청.body)});
+      
+      if(session && target.author === session.user.email) {
+        let result = await db.collection("post").deleteOne({
+          _id: new ObjectId(요청.body),
+        });
+        응답.status(200).json("삭제완료");
+      } else {
+        응답.status(500).json("작성자가 아닙니다.");
+      }
+      // console.log(result); // { acknowledged: true, deletedCount: 1 }
     } catch (error) {
       console.log(error.message);
     }
